added worker class
