+++ zapalki/src/zapalki/BlueMatch.java
@@ -1,7 +1,5 @@
 package zapalki;
 
-import gui.GuiManager;
-
 public class BlueMatch extends Match{
     BlueMatch(){
         color = "blue";
