+++ zapalki/src/zapalki/Command.java
@@ -1,4 +1,6 @@
-package console_input;
+package zapalki;
+
+import zapalki.Items;
 
 public interface Command {
     void execute(Items data);
