+++ zapalki/src/zapalki/Items.java
@@ -1,20 +1,13 @@
-package console_input;
-
-import resources.Resources;
-import zapalki.Match;
-import zapalki.MatchBox;
+package zapalki;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Items {
-    private final int wood = 0;
-    private final int sulfur = 0;
-    private final int paper = 0;
-
-    private Resources resources = new Resources(this.wood, this.sulfur, this.paper);
+    private Resources resources = new Resources(0, 0, 0);
     private List<MatchBox> boxes = new ArrayList<>();
     private List<Match> matches = new ArrayList<>();
+    private List<Worker> workers = new ArrayList<>();
 
     public Resources getResources() {
         return resources;
@@ -29,8 +22,12 @@ public class Items {
     }
 
     public void reset() {
-        this.resources = new Resources(wood, sulfur, paper);
+        this.resources = new Resources(0,0,0);
         this.boxes = new ArrayList<>();
         this.matches = new ArrayList<>();
+    }
+
+    public List<Worker> getWorkers() {
+        return workers;
     }
 }
