+++ zapalki/src/zapalki/MatchBox.java
@@ -1,7 +1,5 @@
 package zapalki;
 
-import gui.GuiManager;
-
 import java.util.List;
 import java.util.stream.Collectors;
