+++ zapalki/src/zapalki/MatchFactory.java
@@ -1,7 +1,5 @@
 package zapalki;
 
-import gui.GuiManager;
-
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
