+++ zapalki/src/zapalki/Resources.java
@@ -1,23 +1,28 @@
-package resources;
-
-import gui.GuiManager;
+package zapalki;
 
 public class Resources {
     private int wood;
     private int sulfur;
     private int paper;
+    private int money;
 
     public Resources(int wood, int sulfur, int paper) {
+        this(wood,sulfur,paper,0);
+    }
+
+    public Resources(int wood, int sulfur, int paper, int money) {
         this.wood = wood;
         this.sulfur = sulfur;
         this.paper = paper;
+        this.money = money;
     }
 
     public static Resources fromString(String type, int amount) {
         switch (type) {
-        case "wood": return new Resources(amount, 0, 0);
-        case "sulfur": return new Resources(0, amount, 0);
-        case "paper": return new Resources(0, 0, amount);
+        case "wood": return new Resources(amount, 0, 0,0);
+        case "sulfur": return new Resources(0, amount, 0,0);
+        case "paper": return new Resources(0, 0, amount,0);
+        case "money": return new Resources(0, 0, 0,amount);
         default:
             GuiManager.print("Invalid resource type");
             return null;
@@ -33,12 +38,14 @@ public class Resources {
         wood += amount.wood;
         sulfur += amount.sulfur;
         paper += amount.paper;
+        money += amount.money;
     }
 
     public void subtract(Resources amount) {
         wood -= amount.wood;
         sulfur -= amount.sulfur;
         paper -= amount.paper;
+        money -= amount.money;
     }
 
     @Override
@@ -47,14 +54,15 @@ public class Resources {
                 "wood=" + wood +
                 ", sulfur=" + sulfur +
                 ", paper=" + paper +
+                ", money=" + money +
                 '}';
     }
 
     public boolean isEnough(Resources r){
-        return (this.wood >= r.wood && this.sulfur >= r.sulfur && this.paper >= r.paper);
+        return (this.wood >= r.wood && this.sulfur >= r.sulfur && this.paper >= r.paper && this.money >= r.money);
     }
 
     public static Resources multiply(Resources r, int mul){
-        return new Resources(r.wood*mul,r.sulfur*mul,r.paper*mul);
+        return new Resources(r.wood*mul,r.sulfur*mul,r.paper*mul,r.money * mul);
     }
 }
