+++ zapalki/src/zapalki/YellowMatch.java
@@ -1,7 +1,5 @@
 package zapalki;
 
-import gui.GuiManager;
-
 public class YellowMatch extends Match{
     YellowMatch(){
         color = "yellow";
